requirements.txt
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
platformdirs==3.2.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
rsa==4.9
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
yarg==0.1.9
yarl==1.8.2
